Look for TTF fonts in program directory and default font path too

+++ src/gui/sdl_ttf.c
@@ -385,7 +385,7 @@ TTF_Font* TTF_OpenFontIndexRW( SDL_RWops *src, int freesrc, int ptsize, long ind
 	}
 
 	/* Check to make sure we can seek in this stream */
-	position = SDL_RWtell(src);
+	position = (int)SDL_RWtell(src);
 	if ( position < 0 ) {
 		TTF_SetError( "Can't seek in stream" );
 		return NULL;
